Fixed final few tests
