+++ tests/test_algorithms/test_llms/conftest.py
@@ -6,9 +6,7 @@ from importlib.util import find_spec
 import numpy as np
 import pytest
 import torch
-from accelerate import Accelerator
 from accelerate.state import AcceleratorState
-from accelerate.utils import DeepSpeedPlugin
 from agilerl.utils.ppo_value_head import AutoModelForCausalLMWithValueHead
 
 from tests.utils import (
@@ -66,37 +64,6 @@ def set_seed():
         torch.cuda.manual_seed_all(SEED)
     np.random.seed(SEED)
     random.seed(SEED)
-
-
-def generate_accelerator(use_deepspeed_optimizer, config):
-    if config is not None and not torch.cuda.is_available():
-        pytest.skip("DeepSpeed-configured LLM tests require CUDA support.")
-    if config is not None and find_spec("deepspeed") is None:
-        pytest.skip("DeepSpeed-configured LLM tests require deepspeed.")
-
-    gc.collect()
-    torch.cuda.empty_cache()
-    AcceleratorState._reset_state(True)
-    if use_deepspeed_optimizer and (config is not None):
-        config["optimizer"] = {
-            "type": "AdamW",
-            "params": {
-                "lr": 1e-4,  # Smaller learning rate
-                "betas": [0.9, 0.999],
-                "eps": 1e-8,
-                "weight_decay": 0.01,
-            },
-        }
-    return (
-        Accelerator(deepspeed_plugin=DeepSpeedPlugin(hf_ds_config=config))
-        if config is not None
-        else None
-    )
-
-
-@pytest.fixture(scope="function")
-def accelerator_factory():
-    return generate_accelerator
 
 
 def generate_model(pretrained_model_name_or_path, add_value_head=False):
